requirements.txt
# requirements.txt

# Core dependencies
tensorflow>=2.10.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0

# Image processing
opencv-python>=4.5.0
Pillow>=9.0.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
joblib>=1.1.0

# Development
jupyter>=1.0.0
ipykernel>=6.0.0
black>=22.0.0
flake8>=4.0.0

# Web Backend
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# --- NEW: Database & Security ---
# SQLAlchemy is our translator between Python and Postgres
sqlalchemy>=2.0.0
# psycopg2-binary is the physical driver for PostgreSQL
psycopg2-binary>=2.9.0
# python-dotenv for managing your database password securely
python-dotenv>=1.0.0
# alembic for database migrations (optional but recommended)
alembic>=1.11.0